[TensorPipe] Do not require user to provide worker name-to-rank map
Pull Request resolved: https://github.com/pytorch/pytorch/pull/38052

The initial version of the TensorPipe agent required the user to specify the full map between workers' names and their ids, on each worker. However it's enough for each worker to just specify their name and id, as these can then be exchanged using the store.

Addresses #37784, although I think we can go further and use the store to also automatically assign ranks to workers, so that the user only needs to specify a name.
ghstack-source-id: 103741372

Differential Revision: [D21463833](https://our.internmc.facebook.com/intern/diff/D21463833/)